Filtration Interface

Ripserer.AbstractFiltration — Type

AbstractFiltration{I, T}

A filtration is used to find the edges in filtration and to create simplices. An AbstractFiltration{I, T}'s simplex type is expected to return simplices of type <:AbstractSimplex{_, T, I}.

Interface

n_vertices(::AbstractFiltration)

edges(::AbstractFiltration)

simplex_type(::AbstractFiltration, dim)

simplex(::AbstractFiltration, ::Val{dim}, vertices, sign)

unsafe_simplex(::AbstractFiltration, ::Val{dim}, vertices, sign)

unsafe_cofacet(::AbstractFiltration, simplex, vertices, vertex[, sign, edges])

birth(::AbstractFiltration, v)

threshold(::AbstractFiltration)

columns_to_reduce(::AbstractFiltration)

emergent_pairs(::AbstractFiltration)

postprocess_diagram(::AbstractFiltration, ::Any)

Ripserer.n_vertices — Method

n_vertices(::AbstractFiltration)

Return the number of vertices in filtration.

Ripserer.edges — Method

edges(::AbstractFiltration)

Get edges (1-simplices) in filtration. Edges should be of type simplex_type(filtration, 1).

PersistenceDiagrams.birth — Method

birth(::AbstractFiltration, v)
birth(::AbstractFiltration)

Get the birth time of vertex v in filtration. Defaults to 0. When v is not given, return births in an array of same size as vertices(::AbstractFiltration).

PersistenceDiagrams.threshold — Method

threshold(::AbstractFiltration)

Get the threshold of filtration. This is the maximum diameter a simplex in the filtration can have. Defaults to Inf.

Ripserer.simplex_type — Function

simplex_type(::Type{<:AbstractFiltration}, D)
simplex_type(::AbstractFiltration, D)

Return the D-dimensional simplex type in the filtration. Only the method for the type needs to be overloaded.

Ripserer.simplex — Function

 simplex(::AbstractFiltration, ::Val{D}, vertices, sign=1)

Return D-simplex constructed from vertices with sign equal to sign. Return nothing if simplex is not in filtration. This function is safe to call with vertices that are out of order. Default implementation sorts vertices and calls unsafe_simplex.

Ripserer.unsafe_simplex — Function

unsafe_simplex(::AbstractFiltration, ::Val{D}, vertices, sign=1)

Return D-simplex constructed from vertices with sign equal to sign. Return nothing if simplex is not in filtration. The unsafe in the name implies that it's up to the caller to ensure vertices are sorted and unique.

Ripserer.unsafe_cofacet — Function

unsafe_cofacet(filtration, simplex, cofacet_vertices, v, sign[, edges])
unsafe_cofacet(::Type{S}, filtration, simplex, cofacet_vertices, v, sign[, edges])

Return cofacet of simplex with vertices equal to cofacet_vertices. v is the vertex that was added to construct the cofacet. In the case of sparse rips filtrations, an additional argument edges is used. edges is a vector that contains the weights on edges connecting the new vertex to old vertices. S is the simplex type which can be used for dispatch.

The unsafe in the name implies that it's up to the caller to ensure vertices are sorted and unique.

Default implementation uses unsafe_simplex.

Ripserer.columns_to_reduce — Function

columns_to_reduce(::AbstractFilration, prev_column_itr)

List all columns to reduce in next dimension, possibly computing it from previous columns. Default implementation uses coboundary with the all cofacets parameter set to Val(false).

Ripserer.emergent_pairs — Function

emergent_pairs(::AbstractFiltration)

Perform the emergent pairs optimization. Default to returning true. Should be set to false for a filtration type that is unable to produce (co)boundary simplices in the correct order.

Ripserer.postprocess_diagram — Function

postprocess_diagram(::AbstractFiltration, diagram)

This function is called on each resulting persistence diagram after all intervals have been computed. Defaults to sorting the diagram.

Ripserer.AbstractRipsFiltration — Type

AbstractRipsFiltration{I<:Signed, T} <: AbstractFiltration{I, T}

An abstract Vietoris-Rips filtration. Its subtypes can only overload dist and get default implementations for the rest of the filtration interface.

Ripserer.dist — Function

dist(::AbstractRipsFiltration, u, v)
dist(::AbstractRipsFiltration)

Return the distance between vertices u and v. If the distance is somehow invalid, it may return missing instead. If u and v are not given, return the distance matrix.

Simplex Interface

Ripserer.AbstractSimplex — Type

AbstractSimplex{D, T, I} <: AbstractVector{I}

An abstract type for representing simplices. A simplex must have a diameter of type T, which is its birth time. The dimension must be encoded in the type as D and can be accessed by dim.

The simplex is expected to act like an array of indices of type I, but this is not actually needed for the main algorithm.

Interface

AbstractSimplex{D}(::SVector{<:Any, <:Integer}, ::T)

birth(::AbstractSimplex)

index(::AbstractSimplex)

sign(::AbstractSimplex)

coboundary(::Any, ::AbstractSimplex)

boundary(::Any, ::AbstractSimplex)

vertices(::AbstractSimplex)

Base.:-(::AbstractSimplex)

PersistenceDiagrams.dim — Method

dim(::AbstractSimplex)
dim(::Type{<:AbstractSimplex})

Get the dimension of simplex i.e. the value of D.

dim(Simplex{2}((3, 2, 1), 3.2))

# output

2

PersistenceDiagrams.birth — Method

birth(simplex::AbstractSimplex)

Get the birth time of simplex, i.e. the time it appears in a filtration.

birth(Simplex{2}((3, 2, 1), 3.2))

# output

3.2

Ripserer.index — Method

index(simplex::AbstractSimplex)

Get the combinatorial index of the simplex. The index can be any type, but should uniquely identify a simplex. It is also used to break ties when comparing simplices with the same birth time.

index(Simplex{2}((3, 2, 1), 3.2))

# output

1

Base.sign — Method

sign(simplex::AbstractSimplex)

Get the orientation of simplex. Should return -1 or 1.

sign(Simplex{2}((3, 2, 1), 3.2))

# output

+1

Base.:- — Method

-(simplex::AbstractSimplex)

Reverse the simplex orientation.

-Simplex{2}((3, 2, 1), 3.2)

# output

2-dim Simplex(1, 1):
  -[3, 2, 1]

Ripserer.vertices — Method

vertices(simplex::AbstractSimplex{dim, T, I})

Get the vertices of simplex. Returns SVector{length(simplex), I}. When index(simplex) is an interger, a default implementation is provided.

vertices(Simplex{2}((3, 2, 1), 3.2))

# output

3-element StaticArrays.SArray{Tuple{3},Int64,1,3} with indices SOneTo(3):
 3
 2
 1

Ripserer.coboundary — Method

coboundary(filtration, simplex[, Val{all_cofacets}])

Iterate over the coboundary of simplex by decreasing index. Use the filtration to determine the diameters and validity of cofacets.

If all_cofacets is false, only return cofaces with vertices added to the beginning of vertex list. The method with all_cofacets only has to be implemented if the filtration does not overload columns_to_reduce.

Comes with a default implementation.

Warning

If cofacets are not returned in decreasing index, the algorithm will not work correctly.

Example

filtration = Rips([0 1 1 1; 1 0 1 1; 1 1 0 1; 1 1 1 0])

for c in coboundary(filtration, Simplex{1}(2, 1))
    println(c)
end

# output

Simplex{2}(+[4, 3, 1], 1)
Simplex{2}(-[3, 2, 1], 1)

for c in coboundary(filtration, Simplex{1}(2, 1), Val(false))
    println(c)
end

# output

2-dim Simplex(1, 1):
  +[4, 3, 1]

Ripserer.boundary — Method

boundary(filtration, simplex[, Val{all_cofacets}])

Iterate over the boundary of simplex by increasing index. Use the filtration to determine the diameters and validity of cofacets.

Comes with a default implementation.

filtration = Rips([0 1 1 1; 1 0 1 1; 1 1 0 1; 1 1 1 0])

for f in boundary(filtration, Simplex{2}(2, 1))
    println(f)
end

# output

Simplex{2}(+[2, 1], 1)
Simplex{2}(-[4, 1], 1)
Simplex{2}(+[4, 2], 1)