Filtration Interface

Ripserer.AbstractFiltration — Type

AbstractFiltration

A filtration is used to find the edges in filtration and to determine diameters of simplices.

Interface

n_vertices(::AbstractFiltration)

edges(::AbstractFiltration)

simplex_type(::AbstractFiltration, dim)

simplex(::AbstractFiltration, ::Val{dim}, vertices, sign)

unsafe_simplex(::AbstractFiltration, ::Val{dim}, vertices, sign)

unsafe_cofacet(::AbstractFiltration, simplex, vertices, vertex[, edges, sign])

birth(::AbstractFiltration, v)

threshold(::AbstractFiltration)

postprocess_interval(::AbstractFiltration, ::Any)

Ripserer.n_vertices — Method

n_vertices(filtration::AbstractFiltration)

Return the number of vertices in filtration.

Ripserer.edges — Method

edges(filtration::AbstractFiltration)

Get edges (1-simplices) in filtration. Edges should be of type simplex_type(filtration, 1).

PersistenceDiagrams.birth — Method

birth(::AbstractFiltration, v)

Get the birth time of vertex v in filtration. Defaults to 0.

PersistenceDiagrams.threshold — Method

threshold(::AbstractFiltration)

Get the threshold of filtration. This is the maximum diameter a simplex in the filtration can have. Defaults to Inf.

Ripserer.simplex_type — Function

simplex_type(::AbstractFiltration, d)

Return the d-dimensional simplex type in the filtration.

Ripserer.simplex — Function

 simplex(::AbstractFiltration, ::Val{D}, vertices, sign=1)

Return D-simplex constructed from vertices with sign equal to sign. Return nothing if simplex is not in filtration. This function is safe to call with vertices that are out of order. Default implementation sorts vertices and calls unsafe_simplex.

Ripserer.unsafe_simplex — Function

unsafe_simplex(::AbstractFiltration, ::Val{D}, vertices, sign=1)

Return D-simplex constructed from vertices with sign equal to sign. Return nothing if simplex is not in filtration. The unsafe in the name implies that it's up to the caller to ensure vertices are sorted and unique.

Ripserer.unsafe_cofacet — Function

unsafe_cofacet(filtration, simplex, cofacet_vertices, new_vertex[, edges, sign=1])

Return cofacet of simplex with vertices equal to cofacet_vertices. new_vertex is the vertex that was added to construct the cofacet. In the case of sparse rips filtrations, an additional argument edges is used. edges is a vector that contains the weights on edges connecting the new vertex to old vertices.

The unsafe in the name implies that it's up to the caller to ensure vertices are sorted and unique.

Default implementation uses unsafe_simplex.

Ripserer.postprocess_interval — Function

postprocess_interval(::AbstractFiltration, interval)

This function is called on each resulting persistence interval. The default implementation does nothing.

Ripserer.AbstractRipsFiltration — Type

AbstractRipsFiltration{I<:Signed, T} <: AbstractFiltration

An abstract Vietoris-Rips filtration. Its subtypes can overload dist and get the following default implementations.

edges

simplex_type.

simplex.

unsafe_simplex.

unsafe_cofacet.

Ripserer.dist — Function

dist(::AbstractRipsFiltration, u, v)
dist(::AbstractRipsFiltration)

Return the distance between vertices u and v. If the distance is somehow invalid, it may return missing instead. If u and v are not given, return the distance matrix.