API

Ripserer.ripserer — Function

ripserer(dists::AbstractMatrix{T}; dim_max=1, modulus=2, threshold=typemax(T))

Compute the persistent homology of metric space represented by dists.

Keyoword Arguments

dim_max: compute persistent homology up to this dimension.

modulus: compute persistent homology with coefficients in the prime field of integers mod modulus.

threshold: compute persistent homology up to diameter smaller than threshold. Defaults to radius of input space.

ripserer(filtration::AbstractFiltration)

Compute persistent homology from filtration object.

Ripserer.RipsFiltration — Type

RipsFiltration{T, S<:AbstractSimplex{<:Any, T}}

This type holds the information about the input values. The distance matrix has to be a dense matrix.

Constructor

RipsFiltration(distance_matrix;
               dim_max=1,
               modulus=2,
               threshold=default_threshold(dist),
               eltype=Simplex{modulus, T})

Ripserer.SparseRipsFiltration — Type

SparseRipsFiltration{T, S<:AbstractSimplex{<:Any, T}}

This type holds the information about the input values. The distance matrix will be converted to a sparse matrix with all values greater than threshold deleted. Off-diagonal zeros in the matrix are treaded as typemax(T).

Constructor

SparseRipsFiltration(distance_matrix;
                     dim_max=1,
                     modulus=2,
                     threshold=default_threshold(dist),
                     eltype=Simplex{modulus, T})

Adding New Simplex Types

Ripserer.AbstractSimplex — Type

AbstractSimplex{C, T}

An abstract type for representing simplices. A simplex is represented by its diameter, combinatorial index and coefficient value. It does not need to hold information about its dimension or the vertices it includes.

T is the type of distance and C is the coefficient type.

Interface

index(::AbstractSimplex)

coef(::AbstractSimplex)

set_coef(::AbstractSimplex{C}, ::C)

diam(::AbstractSimplex)

Ripserer.index — Method

index(simplex::AbstractSimplex)

Get the combinatorial index of simplex. The index of is equal to

\[(i_d, i_{d-1}, ..., 1) \mapsto \sum_{k=1}^{d+1} \binom{i_k - 1}{k},\]

where i_k are the simplex vertex indices.

Ripserer.coef — Function

coef(simplex::AbstractSimplex)

Get the coefficient value of simplex.

Ripserer.set_coef — Function

set_coef(simplex::AbstractSimplex, value)

Return new simplex with new coefficient value.

Ripserer.diam — Method

diam(simplex::AbstractSimplex)

Get the diameter of simplex.

Adding New Filtration Types

Ripserer.AbstractFiltration — Type

AbstractFiltration{T, S<:AbstractSimplex{C, T}}

An abstract type that holds information about the distances between vertices and the simplex type.

Interface

Base.length(::AbstractFiltration)

dist(::AbstractFiltration, ::Integer, ::Integer)

edges(::AbstractFiltration)

dim_max(::AbstractFiltration)

diam(::AbstractFiltration, iterable) - optional, defaults to diameter of vertex set.

Base.binomial(::AbstractFiltration, n, k) - optional, but recommended.

threshold(::AbstractFiltration) - optional, defaults to typemax(T).

Base.length — Method

length(filtration::AbstractFiltration)

Number of vertices in filtration.

Ripserer.dist — Function

dist(filtration::AbstractFiltration, i, j)

Get the distance between vertex i and vertex j.

Ripserer.edges — Function

edges(filtration::AbstractFiltration)

Get edges in distance matrix in filtration, sorted by decresing length and increasing combinatorial index.

Ripserer.dim_max — Function

dim_max(filtration::AbstractFiltration)

Get the maximum dimension of simplices in filtration.

Ripserer.diam — Method

diam(flt::AbstractFiltration, vertices)
diam(flt::AbstractFiltration, vertices, vertex)

Get the diameter of list of vertices i.e. diameter of simplex with vertices. If additional vertex is given, only calculate max distance from vertices to vertex.

Base.binomial — Method

binomial(filtration::AbstractFiltration, n, k)

An abstract filtration may have binomial coefficients precomputed for better performance.

Ripserer.threshold — Function

threshold(flt::AbstractFiltration)

Get the threshold of flt. Simplices with diameter strictly larger than this value will be ignored.

SparseArrays.issparse — Method

SparseArrays.issparse(::Type{A}) where A<:AbstractFiltration

Return true if A is a sparse filtration. A filtration should be sparse if most simplices are to be skipped. Defaults to false.