API

Ripserer.ripserer — Function

ripserer(dists::AbstractMatrix{T}; dim_max=1, modulus=2, threshold=typemax(T))
ripserer(points, metric; dim_max=1, modulus=2, threshold=typemax(T))

Compute the persistent homology of metric space represented by dists or points and metric.

Keyoword Arguments

dim_max: compute persistent homology up to this dimension.

modulus: compute persistent homology with coefficients in the prime field of integers mod modulus.

threshold: compute persistent homology up to diameter smaller than threshold. Defaults to radius of input space.

sparse: if true, use SparseRipsFiltration. Defaults to issparse(dists).

ratio: only keep intervals with death(interval) > birth(interval) * ratio. Defaults to 1.

ripserer(filtration::AbstractFiltration; dim_max=1)

Compute persistent homology from filtration object.

Ripserer.RipsFiltration — Type

RipsFiltration{T, S<:AbstractSimplex{<:Any, T}} <: AbstractFlagFiltration{T, S}

Constructor

RipsFiltration(
    distance_matrix;
    modulus=2,
    threshold=default_rips_threshold(dist),
    edge_type=Simplex{1, modulus, T, Int, UInt}
)

Ripserer.SparseRipsFiltration — Type

SparseRipsFiltration{T, S<:AbstractSimplex{<:Any, T}} <: AbstractFlagFiltration{T, S}

This type holds the information about the input values. The distance matrix will be converted to a sparse matrix with all values greater than threshold deleted. Off-diagonal zeros in the matrix are treated as ∞.

Constructor

SparseRipsFiltration(
    distance_matrix;
    modulus=2,
    threshold=default_rips_threshold(dist),
    eltype=Simplex{modulus, T}
)

Simplex Types

Ripserer.AbstractSimplex — Type

AbstractSimplex{D, C, T}

An abstract type for representing simplices. A simplex is represented by its dimension, diameter, combinatorial index and coefficient value. It does not need to hold information about its the vertices it includes, since they can be recomputed from the index and dimension.

D is the dimension, T is the type of distance and C is the coefficient type.

Interface

index(::AbstractSimplex)

coef(::AbstractSimplex)

set_coef(::AbstractSimplex, ::Any)

diam(::AbstractSimplex)

vertices(::AbstractSimplex)

Ripserer.index — Method

index(simplex::AbstractSimplex)

Get the combinatorial index of simplex.

Ripserer.coef — Function

coef(simplex::AbstractSimplex)

Get the coefficient value of simplex.

Ripserer.set_coef — Function

set_coef(simplex::AbstractSimplex, value)

Return new simplex with new coefficient value.

Ripserer.diam — Method

diam(simplex::AbstractSimplex)

Get the diameter of simplex.

Ripserer.dim — Function

dim(::AbstractSimplex)
dim(::Type{AbstractSimplex})

Get the dimension of simplex i.e. the value of D.

dim(::PersistenceDiagram)

Get the dimension of persistence diagram.

Ripserer.diam — Function

diam(simplex::AbstractSimplex)

Get the diameter of simplex.

diam(flt::AbstractFiltration, vertices)

Get the diameter of list of vertices i.e. diameter of simplex with vertices.

diam(flt::AbstractFiltration, simplex, vertices, vertex)

Get the diameter of coface of simplex that is formed by adding vertex to vertices.

Filtration Types

Ripserer.AbstractFiltration — Type

AbstractFiltration{T, S<:AbstractSimplex{C, T}}

An abstract type that holds information about the distances between vertices and the simplex type.

Interface

n_vertices(::AbstractFiltration) - return number of vertices in filtration.

edges(::AbstractFiltration) - return all edges in filtration as (l, (i, j)) where l is the edge length and i and j are its endpoints.

diam(::AbstractFiltration, vs) - diameter of simplex with vertices in vs. Should return Infinity() if simplex is above threshold.

diam(::AbstractFiltration, sx::AbstractSimplex, vs, u) - diameter of simplex sx with vertices in vs and an added vertex u. Should return Infinity() if simplex is above threshold.

SparseArrays.issparse(::Type{A}) where A<:AbstractFiltration - optional, defaults to false. Should be true if most of the simplices are expected to be skipped.

Ripserer.AbstractFlagFiltration — Type

AbstractFlagFiltration{T, S} <: AbstractFiltration{T, S}

An abstract flag filtration is a filtration of flag complexes. Its subtypes can overload dist(::AbstractFlagFiltration{T}, u, v)::Union{T, Infinity} instead of diam. diam(::AbstractFlagFiltration, ...) defaults to maximum dist among vertices.

Ripserer.n_vertices — Function

n_vertices(filtration::AbstractFiltration)

Number of vertices in filtration.

Ripserer.edges — Function

edges(filtration::AbstractFiltration)

Get edges in distance matrix in filtration, sorted by decresing length and increasing combinatorial index. Edges should be of type edge_type(filtration).

edges(dist::AbstractMatrix{T}, thresh, S)

Return sorted edges of type S in distance matrix with length lower than thresh.

Ripserer.diam — Method

diam(flt::AbstractFiltration, vertices)

Get the diameter of list of vertices i.e. diameter of simplex with vertices.

Ripserer.diam — Method

diam(flt::AbstractFiltration, simplex, vertices, vertex)

Get the diameter of coface of simplex that is formed by adding vertex to vertices.

SparseArrays.issparse — Method

SparseArrays.issparse(::Type{A}) where A<:AbstractFiltration

Return true if A is a sparse filtration. A filtration should be sparse if most simplices are to be skipped. Defaults to false.