API

Ripserer

Ripserer.ripserer — Function

ripserer(dists::AbstractMatrix{T}; kwargs...)
ripserer(points; metric=Euclidean(), births, kwargs...)

Compute the persistent homology of metric space represented by dists or points and metric. points must be an array of bitstypes, such as NTuples or SVectors.

Keyoword Arguments

dim_max: compute persistent homology up to this dimension. Defaults to 1.

modulus: compute persistent homology with coefficients in the prime field of integers mod modulus. Defaults to 2.

threshold: compute persistent homology up to diameter smaller than threshold. For Rips filtrations, it defaults to radius of input space.

sparse: if true, use SparseRipsFiltration. Defaults to false || issparse(dists).

ratio: only keep intervals with death(interval) > birth(interval) * ratio. Defaults to 1.

representatives: if true, return representative cocycles along with persistence intervals. Defaults to false.

metric: when calculating persistent homology from points, any metric from Distances.jl can be used. Defaults to Euclidean().

births: when calculating persistent homology from points, births can be used to add birth times to vertices. Defaults to all births equal to 0.

ripserer(filtration::AbstractFiltration; dim_max=1)

Compute persistent homology from filtration object.

Ripserer.RipsFiltration — Type

RipsFiltration{T, V<:AbstractSimplex{<:Any, T}} <: AbstractFlagFiltration{T, V}

This type represents a filtration of Vietoris-Rips complexes. Diagonal items are treated as vertex birth times.

Constructor

RipsFiltration(
    distance_matrix;
    modulus=2,
    threshold=default_rips_threshold(dist),
    vertex_type=Simplex{0, modulus, T, Int, UInt}
)

Ripserer.SparseRipsFiltration — Type

SparseRipsFiltration{T, V<:AbstractSimplex{<:Any, T}} <: AbstractFlagFiltration{T, V}

This type represents a filtration of Vietoris-Rips complexes. The distance matrix will be converted to a sparse matrix with all values greater than threshold deleted. Off-diagonal zeros in the matrix are treated as ∞. Diagonal items are treated as vertex birth times.

Constructor

SparseRipsFiltration(
    distance_matrix;
    modulus=2,
    threshold=nothing,
    eltype=Simplex{modulus, T}
)

Persistence Intervals and Diagrams

Ripserer.PersistenceInterval — Type

PersistenceInterval{T, C}

The type that represents a persistence interval. It behaves exactly like a Tuple{T, Union{T, Infinity}}, but may have a representative cocycle attached to it.

Ripserer.PersistenceDiagram — Type

PersistenceDiagram{P<:PersistenceInterval} <: AbstractVector{P}

Type for representing persistence diagrams. Behaves exactly like an array of PersistenceIntervals, but is aware of its dimension and supports pretty printing and plotting.

Missing docstring.

Missing docstring for RecipesBase.plot(::Union{PersistenceDiagram, AbstractVector{<:PersistenceDiagram}}). Check Documenter's build log for details.

Ripserer.barcode — Method

barcode(diagram; infinity=nothing)

Plot the barcode plot or AbstractVector of diagrams. The infinity keyword argument determines where the infinity line is placed. If set to nothing the function tries to guess a good infinity poistion.

Simplex Types

Ripserer.AbstractSimplex — Type

AbstractSimplex{D, C, T}

An abstract type for representing simplices. A simplex is represented by its dimension, diameter, combinatorial index and coefficient value. It does not need to hold information about its the vertices it includes, since they can be recomputed from the index and dimension.

D is the dimension, T is the type of distance and C is the coefficient type. D is accessible by dim(::AbstractSimplex).

Interface

index(::AbstractSimplex)

coef(::AbstractSimplex)

set_coef(::AbstractSimplex, ::Any)

diam(::AbstractSimplex)

coface_type(::AbstractSimplex)

vertices(::AbstractSimplex) - optional, comes with a default implementation.

coboundary(::AbstractFiltration, ::AbstractSimplex) - optional, comes with a default implementation.

Ripserer.index — Method

index(simplex::AbstractSimplex)

Get the combinatorial index of simplex.

Ripserer.coef — Method

coef(simplex::AbstractSimplex)

Get the coefficient value of simplex.

Ripserer.set_coef — Method

set_coef(simplex::AbstractSimplex, value)

Return new simplex of the same type with new coefficient value.

Ripserer.diam — Method

diam(simplex::AbstractSimplex)

Get the diameter of simplex.

Ripserer.dim — Method

dim(::AbstractSimplex)
dim(::Type{AbstractSimplex})

Get the dimension of simplex i.e. the value of D.

Ripserer.vertices — Method

vertices(index, ::Val{dim})

Get the vertices of simplex represented by index. Returns NTuple{dim+1, Int}.

vertices(sx::AbstractSimplex{dim})

Get the vertices of simplex sx. Returns NTuple{dim+1, Int}.

Ripserer.coface_type — Method

coface_type(::AbstractSimplex)
coface_type(::Type{<:AbstractSimplex})

Get the type of coface a simplex hax. For a D-dimensional simplex, this is usually its D+1-dimensional counterpart. Only the coface_type(::Type{<:AbstractSimplex}) method needs to be implemented.

Ripserer.coboundary — Function

coboundary(filtration, simplex)

Find the coboundary of simplex. Use the filtration to determine the diameters and validity of cofaces. Iterates values of the type coface_type(simplex).

Filtration Types

Ripserer.AbstractFiltration — Type

AbstractFiltration{T, V<:AbstractSimplex{0, C, T}}

A filtration is used to find the edges in filtration and to determine diameters of simplices.

T is the distance type, accessible by dist_type and S is the edge type, accesible by edge_type.

Interface

n_vertices(::AbstractFiltration)

edges(::AbstractFiltration)

diam(::AbstractFiltration, vs)

diam(::AbstractFiltration, ::AbstractSimplex, ::Any, ::Any)

SparseArrays.issparse(::Type{A}) where A<:AbstractFiltration - optional, defaults to false.

[birth(::AbstractFiltration, v)] - optional, defaults to returning zero(T).

Ripserer.AbstractFlagFiltration — Type

AbstractFlagFiltration{T, V} <: AbstractFiltration{T, V}

An abstract flag filtration is a filtration of flag complexes. Its subtypes can overload dist(::AbstractFlagFiltration{T}, u, v)::Union{T, Infinity} instead of diam. diam(::AbstractFlagFiltration, ...) defaults to maximum dist among vertices.

Ripserer.n_vertices — Function

n_vertices(filtration::AbstractFiltration)

Number of vertices in filtration.

Ripserer.edges — Function

edges(filtration::AbstractFiltration)

Get edges in distance matrix in filtration, sorted by decresing length and increasing combinatorial index. Edges should be of type edge_type(filtration).

edges(dist::AbstractMatrix{T}, thresh, E)

Return sorted edges of type E in distance matrix with length lower than thresh.

Ripserer.diam — Method

diam(flt::AbstractFiltration, vertices)

Get the diameter of list of vertices i.e. diameter of simplex with vertices.

Missing docstring.

Missing docstring for diam(::AbstractFiltration, ::Any, ::Any, ::Any). Check Documenter's build log for details.

SparseArrays.issparse — Method

SparseArrays.issparse(::Type{A}) where A<:AbstractFiltration

Return true if A is a sparse filtration. A filtration should be sparse if most simplices are to be skipped. Defaults to false.