Filtration Interface

Ripserer.AbstractFiltration — Type

AbstractFiltration

A filtration is used to find the edges in filtration and to determine diameters of simplices.

Interface

n_vertices(::AbstractFiltration)

edges(::AbstractFiltration)

diam(::AbstractFiltration, vertices)

diam(::AbstractFiltration, ::AbstractSimplex, ::Any, ::Any) - only used when simplex_type is an IndexedSimplex.

simplex_type(::AbstractFiltration, dim)

birth(::AbstractFiltration, v) - optional, defaults to returning zero(T).

threshold(::AbstractFiltration) - optional, defaults to returning Inf.

postprocess_interval(::AbstractFiltration, ::Any) - optional postprocessing function that is applied to each interval in resulting persistence diagram.

Ripserer.n_vertices — Method

n_vertices(filtration::AbstractFiltration)

Return the number of vertices in filtration.

Ripserer.edges — Method

edges(filtration::AbstractFiltration)

Get edges in distance matrix in filtration, sorted by decresing length and increasing combinatorial index. Edges should be of type simplex_type(filtration, 1)`.

Ripserer.diam — Method

diam(::AbstractFiltration, vertices)

Get the diameter of a simplex with the vertex set vertices. Should return missing if vertices do not form a valid simplex.

Ripserer.diam — Method

diam(::AbstractFiltration, simplex, vertices, new_vertex)

Get the diameter of coface of a Simplex that is formed by adding new_vertex to vertices. Should return missing if new simplex is not valid.

This functions is used with the coboundary function for IndexedSimplex

PersistenceDiagrams.birth — Method

birth(::AbstractFiltration, v)

Get the birth time of vertex v in filtration. Defaults to 0.

PersistenceDiagrams.threshold — Method

threshold(::AbstractFiltration)

Get the threshold of filtration. This is the maximum diameter a simplex in the filtration can have. Used only for placing the infinity line in plotting. Defaults to missing.

Ripserer.simplex_type — Function

simplex_type(::AbstractFiltration, d)

Return the d-dimensional simplex type in the filtration.

Ripserer.postprocess_interval — Function

postprocess_diagram(::AbstractFiltration, interval)

This function is called on each resulting persistence interval. The default implementation does nothing.

Ripserer.AbstractRipsFiltration — Type

AbstractRipsFiltration{I<:Signed, T} <: AbstractFiltration

An abstract Vietoris-Rips filtration. Its subtypes can overload dist(::AbstractRipsFiltration{T}, u, v)::Union{T, Missing} instead of diam.

diam(::AbstractRipsFiltration, ...) then defaults to maximum [dist] among vertices.

Comes with default implementations of edges and simplex_type.

Ripserer.dist — Function

dist(::AbstractRipsFiltration, u, v)
dist(::AbstractRipsFiltration)

Return the distance between vertices u and v. If the distance is higher than the threshold, return missing instead. If u and v are not given, return the distance matrix.