Filtrations

Ripserer.AbstractFiltration — Type

AbstractFiltration{T, S<:AbstractSimplex}

A filtration is used to find the edges in filtration and to determine diameters of simplices.

T is the distance type, accessible by dist_type and S is the edge type, accesible by vertex_type.

Interface

n_vertices(::AbstractFiltration)

edges(::AbstractFiltration)

diam(::AbstractFiltration, vertices)

diam(::AbstractFiltration, ::AbstractSimplex, ::Any, ::Any) - only used for Simplex.

birth(::AbstractFiltration, v) - optional, defaults to returning zero(T).

threshold(::AbstractFiltration) - optional, defaults to returning missing.

simplex_type(::AbstractFiltration, dim)

Ripserer.n_vertices — Method

n_vertices(filtration::AbstractFiltration)

Return the number of vertices in filtration.

Ripserer.edges — Method

edges(filtration::AbstractFiltration)

Get edges in distance matrix in filtration, sorted by decresing length and increasing combinatorial index. Edges should be of type edge_type(filtration).

Ripserer.diam — Method

diam(flt::AbstractFiltration, vertices)

Get the diameter of list of vertices i.e. diameter of simplex with vertices. Retrun missing if vertices do not form a valid simplex.

Ripserer.diam — Method

diam(flt::AbstractFiltration, simplex, vertices, vertex)

Get the diameter of coface of a Simplex that is formed by adding vertex to vertices. Return missing if new simplex is not valid.

PersistenceDiagrams.birth — Method

birth(::AbstractFiltration, v)

Get the birth time of vertex v.

PersistenceDiagrams.threshold — Method

threshold(::AbstractFiltration)

Get the threshold of filtration. This is the maximum diameter a simplex in the filtration can have. Used only for placing the infinity line in plotting.

Ripserer.AbstractFlagFiltration — Type

AbstractFlagFiltration{T, S} <: AbstractFiltration{T, S}

An abstract flag filtration is a filtration of flag complexes. Its subtypes can overload dist(::AbstractFlagFiltration{T}, u, v)::Union{T, Missing} instead of diam. diam(::AbstractFlagFiltration, ...) defaults to maximum dist among vertices.