Public API

Ripserer

Ripserer.ripserer — Function

ripserer(dists::AbstractMatrix; kwargs...)
ripserer(points; metric=Distances.Euclidean(), births, kwargs...)
ripserer(filtration::AbstractFiltration; kwargs...)

Compute the persistent homology of metric space represented by dists, points and metric or a Ripserer.AbstractFiltration.

If using points, points must be an array of isbits types, such as NTuples or SVectors.

Keyoword Arguments

dim_max: compute persistent homology up to this dimension. Defaults to 1.

modulus: compute persistent homology with coefficients in the prime field of integers mod modulus. Defaults to 2.

field_type: use this type of field of coefficients. Defaults to Ripserer.Mod{modulus}.

threshold: compute persistent homology up to diameter smaller than threshold. This parameter is only applicable when using distance matrices or points as input. When using filtrations, threshold must be passed to the filtration constructor. Defaults to radius of input space. When using low thresholds with points or distance matrices, consider using SparseRips.

cutoff: only keep intervals with persistence(interval) > cutoff. Defaults to 0.

reps: if true, return representative cocycles along with persistence intervals. Defaults to false.

progress: If true, show a progress bar. Defaults to false.

metric: when calculating persistent homology from points, any metric from Distances.jl can be used. Defaults to Distances.Euclidean().

cohomology: if set to false, compute persistent homology instead of cohomology. This is much slower and gives the same result, but may give more informative representatives when reps is enabled. Currently unable to compute infinite intervals in dimensions higher than 0. Defaults to false.

Ripserer.Rips — Type

Rips{I, T} <: AbstractRipsFiltration{I, T}

This type represents a filtration of Vietoris-Rips complexes. Diagonal items are treated as vertex birth times.

Threshold defaults to radius of input space. When using low thresholds, consider using SparseRips instead.

Constructors

Rips(distance_matrix; threshold=nothing)

Rips(points; metric=Euclidean(), threshold=nothing)

Rips{I}(args...): I sets the size of integer used to represent simplices.

Ripserer.SparseRips — Type

SparseRips{I, T} <: AbstractRipsFiltration{T, Simplex}

This type represents a sparse filtration of Vietoris-Rips complexes. The distance matrix will be converted to a sparse matrix with all values greater than threshold deleted. Off-diagonal zeros in the matrix are treated as missing. Diagonal items are treated as vertex birth times.

Constructor

SparseRips{I}(distance_matrix; threshold=nothing)

SparseRips(distance_matrix; threshold=nothing): I sets the integer size used to represent simplices.

Ripserer.Cubical — Type

Cubical{T, K} <: AbstractFiltration{CartesianIndex{K}, T}

Cubical is used to compute sublevel persistent homology on N-dimensional images, which are of type AbstractArray{T, N}.

This type uses the CubeMap structure to find birth times of cubes (see reference).

Constructor

Cubical(image::AbstractArray{T, N}, threshold=maximum(image))

Reference

Wagner, H., Chen, C., & Vuçini, E. (2012). Efficient computation of persistent homology for cubical data. In Topological methods in data analysis and visualization II (pp. 91-106). Springer, Berlin, Heidelberg.

Ripserer.Custom — Type

Custom{T} <: AbstractFiltration{Int, T}

Build a custom filtration by specifying simplices and their birth times.

The list of simplices is corrected to form a valid filtration; birth times are corrected so a simplex is never born before its faces and missing simplices are added.

See example below for construction. Note how the unlisted 0-simplices were added with birth times equal to the lowest between their cofaces. The order in which simplices are given does not matter.

Note that this is not terribly efficient as Ripserer's implicit algorithm is not optimized for these kinds of filtrations. It can still be useful for experimentation or small filtrations.

Example

```jldoctest julia> flt = Custom([ (1,) => 0, (4,) => 0, (1, 2) => 1, (1, 3) => 2, (1, 4) => 3, (2, 3) => 4, (2, 4) => 5, (3, 4) => 6, (1, 2, 3) => 7, (1, 2, 4) => 8, (1, 3, 4) => 9, ]; threshold=8) Custom{Int64, Int64}(n_vertices=4)

julia> flt[0] # Can be indexed with dimension to list simplices 4-element Array{Simplex{0,Int64,Int64},1}: +Simplex{0}([4], 0) +Simplex{0}([2], 1) +Simplex{0}([3], 2) +Simplex{0}([1], 0)

julia> ripserer(flt)[1] 4-element 0-dimensional PersistenceDiagram: [0.0, 3.0) [1.0, ∞)

julia> ripserer(flt)[2] 3-element 1-dimensional PersistenceDiagram: [4.0, 7.0) [5.0, 8.0) [6.0, ∞)

Ripserer.Mod — Type

Mod{M} <: Integer

Representation of finite field $\mathbb{Z}_M$, integers modulo small, prime M. Supports field arithmetic and can be converted to integer with Int.

Is not comparable on purpose.

Example

Mod{3}(5)

# output

2 mod 3

Mod{3}(5) + 1

# output

0 mod 3

Simplices

Ripserer.Simplex — Type

Simplex{D, T, I<:Integer} <: AbstractSimplex{D, T, I}

The vanilla simplex type represented by dimension D and index of type I and a birth time of type T.

Constructor

Simplex{D[, T, I]}(::I, ::T)

Examples

Simplex{2}(2, 1)

# output

2-dim Simplex{2}(2, 1):
  +[4, 2, 1]

Simplex{10}(Int128(-10), 1.0)

# output

4-dim Simplex{3}(1.0, 10, 2):
  -Int128[12, 11, 10, 9, 8, 7, 6, 5, 4, 2, 1]

Ripserer.Cube — Type

Cube{D, T, K} <: AbstractSimplex{D, T, CartesianIndex{K}}

A Cube is similar to a Simplex, but it has 2^D vertices instead of D+1. The vertices are encoded as the position in the CubeMap.

Ripserer.vertices — Function

vertices(simplex::AbstractSimplex{dim, T, I})

Get the vertices of simplex. Returns SVector{length(simplex), I}. When index(simplex) is an interger, a default implementation is provided.

vertices(Simplex{2}((3, 2, 1), 3.2))

# output

3-element StaticArrays.SArray{Tuple{3},Int64,1,3} with indices SOneTo(3):
 3
 2
 1

vertices(::AbstractFiltration)

Return the vertices in filtration. Defaults to 1:n. The eltype of the result can be anything as long as result[result[i]] == result[i] holds.

PersistenceDiagrams.dim — Function

dim(diagram::PersistenceDiagram)

Get the dimension of persistence diagram. Equivalent to diagram.dim.

dim(::AbstractSimplex)
dim(::Type{<:AbstractSimplex})

Get the dimension of simplex i.e. the value of D.

dim(Simplex{2}((3, 2, 1), 3.2))

# output

2

Persistence Diagrams

See also: PersistenceDiagrams.jl API. For convenience, Ripserer reexports the following:

PersistenceDiagrams.birth

PersistenceDiagrams.death

PersistenceDiagrams.persistence

PersistenceDiagrams.representative

PersistenceDiagrams.birth_simplex

PersistenceDiagrams.death_simplex

PersistenceDiagrams.barcode