Public API

Ripserer

Ripserer.ripserer — Function

ripserer(Type{<:AbstractFiltration}, args...; kwargs...)
ripserer(filtration::AbstractFiltration; kwargs...)

Compute the persistent homology of a filtration. The filtration can be given as an AbstractFiltration type, followed by its arguments, or as an initialized object (see examples below). If only data is given, Rips is used by default.

Keyword Arguments

dim_max: compute persistent homology up to this dimension. Defaults to 1.

modulus: compute persistent homology with coefficients in the prime field of integers mod modulus. Defaults to 2.

field: use this type of field of coefficients. Defaults to Ripserer.Mod{modulus}.

threshold: compute persistent homology up to diameter smaller than threshold. This parameter is only applicable when using distance matrices or points as input. When using filtrations, threshold must be passed to the filtration constructor. Defaults to the radius of the input space. When using low thresholds with points or distance matrices, consider using sparse=true.

cutoff: only keep intervals with persistence(interval) > cutoff. Defaults to 0.

reps: if true, attach representative (co)cycles to persistence intervals. Can also be set to collection of integers to only find representatives in specified dimensions, e.g. reps=1:2 will only find representatives in dimensions 1 and 2. This is useful for large filtrations (such as cubical) where calculating zero-dimensional representatives can be very slow. Defaults to false for cohomology and 1:dim_max for homology.

verbose: If true, show a verbose bar. Defaults to false.

alg: select the algorithm used in computation. The options are:

:cohomology: Default and fastest algorithm. When reps is set, intervals are equipped with representative cocycles.

:homology: Significantly slower than :cohomology, but finds representative cycles. Does not find infinite intervals beyond dimension 0.

:involuted: Use cohomology result to compute representative cycles. Can be extremely efficient compared to :homology, especially with Rips filtrations. Unlike :homology, this algorithm finds infinite intervals.

implicit: If true, an implicit reduction algorithm is used. Defaults to true for :cohomology and :involuted, and false for :homology. implicit=false is not recommended for :cohomology because it disables the emergent pairs optimization.

Other kwargs... are passed to the filtration.

Examples

julia> ts = range(0, 2π; length=20)[1:(end - 1)];

julia> X = [((2 + cos(θ)) * cos(φ), (2 + cos(θ)) * sin(φ), sin(θ)) for θ in ts for φ in ts];

julia> ripserer(X)
2-element Array{PersistenceDiagramsBase.PersistenceDiagram,1}:
 361-element 0-dimensional PersistenceDiagram
 362-element 1-dimensional PersistenceDiagram

julia> ripserer(EdgeCollapsedRips, X; modulus=7, threshold=2)
2-element Array{PersistenceDiagramsBase.PersistenceDiagram,1}:
 361-element 0-dimensional PersistenceDiagram
 362-element 1-dimensional PersistenceDiagram

julia> ripserer(Rips(X; threshold=1); alg=:involuted)
2-element Array{PersistenceDiagramsBase.PersistenceDiagram,1}:
 361-element 0-dimensional PersistenceDiagram
 362-element 1-dimensional PersistenceDiagram

Filtrations

Ripserer.Rips — Type

Rips{I, T} <: AbstractRipsFiltration{I, T}

This type represents a filtration of Vietoris-Rips complexes.

Diagonal items in the input matrix are treated as vertex birth times.

Zero values are not allowed due to how sparse matrices work in Julia. If you need zero birth times, try offseting all values by a constant.

Threshold defaults to the radius of the input space. When using low thresholds, consider using the sparse=true keyword argument. It will give the same result, but may be much faster.

Constructors

Rips(distance_matrix; threshold=nothing)

Rips(points; metric=Euclidean(1e-12), threshold=nothing)

Rips{I}(args...): I sets the size of integer used to represent simplices.

Examples

julia> data = [(sin(t), cos(t)) for t in range(0, 2π, length=101)][1:end-1];

julia> ripserer(Rips(data))
2-element Array{PersistenceDiagramsBase.PersistenceDiagram,1}:
 100-element 0-dimensional PersistenceDiagram
 1-element 1-dimensional PersistenceDiagram

julia> ripserer(Rips(data, threshold=1.7))[2]
1-element 1-dimensional PersistenceDiagram:
 [0.0628, ∞)

julia> using Distances

julia> ripserer(Rips(data, metric=Cityblock()))[2]
1-element 1-dimensional PersistenceDiagram:
 [0.0888, 2.0)

Ripserer.Cubical — Type

Cubical{T, K} <: AbstractFiltration{CartesianIndex{K}, T}

Cubical is used to compute sublevel persistent homology on N-dimensional images, which are of type AbstractArray{T, N}.

This type uses the CubeMap structure to find birth times of cubes (see reference).

Constructor

Cubical(image::AbstractArray{T, N}, threshold=maximum(image))

Reference

Wagner, H., Chen, C., & Vuçini, E. (2012). Efficient computation of persistent homology for cubical data. In Topological methods in data analysis and visualization II (pp. 91-106). Springer, Berlin, Heidelberg.

Example

julia> image = [1 0 0; 0 2 0; 0 0 0];

julia> ripserer(Cubical(image))[1]
1-element 0-dimensional PersistenceDiagram:
 [0.0, ∞)

julia> ripserer(Cubical(image))[2]
1-element 1-dimensional PersistenceDiagram:
 [1.0, 2.0)

Ripserer.Custom — Type

Custom{I, T} <: AbstractCustomFiltration{I, T}

Build a custom filtration by specifying simplices and their birth times.

The list of simplices is corrected to form a valid filtration; birth times are corrected so a simplex is never born before its faces and missing simplices are added.

See the examples below for construction. Note how the unlisted 0-simplices were added with birth times equal to the lowest between their cofaces. The order in which simplices are given does not matter.

To create your own types of custom filtrations, subtype AbstractCustomFiltration.

Examples

julia> flt = Custom([(1,) => 0, (4,) => 0, (1, 2) => 1, (1, 3) => 2, (1, 4) => 3, (2, 3) => 4, (2, 4) => 5, (3, 4) => 6, (1, 2, 3) => 7, (1, 2, 4) => 8, (1, 3, 4) => 9]; threshold=8)
Custom{Int64, Int64}(nv=4)

julia> flt[0] # Can be indexed with dimension to list simplices
4-element Array{Simplex{0,Int64,Int64},1}:
 +Simplex{0}((4,), 0)
 +Simplex{0}((2,), 1)
 +Simplex{0}((3,), 2)
 +Simplex{0}((1,), 0)

julia> ripserer(flt)[1]
2-element 0-dimensional PersistenceDiagram:
 [0.0, 3.0)
 [0.0, ∞)

julia> ripserer(flt)[2]
3-element 1-dimensional PersistenceDiagram:
 [5.0, 8.0)
 [4.0, 7.0)
 [6.0, ∞)

Ripserer.Alpha — Type

Alpha{I, P<:SVector} <: AbstractFiltration{I, Float64}

Alpha filtrations are filtrations of the Delaunay complex.

They have much fewer simplices than Rips, so they are efficient even with large datasets, as long as their dimensionality is low. What "low" means depends on the data, but this is definitely a good choice for 3D or lower. For high dimensional data, filtration construction may take a long time.

Note

Unlike most implementations, this one uses circumdiameters instead of circumradii. This makes the scale of the results comparable to Rips. If you need radius based values, divide your data or the resulting interval endpoints by 2.

Warning

This filtration uses MiniQhull.jl. Please see the installation instructions if constructions cause errors. MiniQhull currently has problems running on Windows. See this issue for more info.

Constructors

Alpha(points; threshold, verbose): points should be a vector of Tuples, SVectors or similar.

Alpha{I}(args...): I sets the size of integer used to represent simplices. Try using I=Int128 if construction complains about overflow.

Reference

Edelsbrunner, H. (1993, July). The union of balls and its dual shape. In Proceedings of the ninth annual symposium on Computational geometry (pp. 218-231).

Example

julia> data = [(sin(t), cos(t), (t - π)^2) for t in range(0, 2π, length=101)[1:end-1]];

julia> alpha = Alpha(data)
Alpha{Int64, Float64}(nv=100)

julia> rips = Rips(data)
Rips{Int64, Float64}(nv=100, sparse=false)

julia> length(Ripserer.edges(alpha))
197

julia> length(Ripserer.edges(rips))
3613

julia> sort(ripserer(alpha)[2], by=persistence)[end]
[0.375, 2.01) with:
 birth_simplex: Ripserer.Simplex{1,Float64,Int64}
 death_simplex: Ripserer.Simplex{2,Float64,Int64}

julia> sort(ripserer(rips)[2], by=persistence)[end]
[0.375, 2.01) with:
 birth_simplex: Ripserer.Simplex{1,Float64,Int64}
 death_simplex: Ripserer.Simplex{2,Float64,Int64}

Ripserer.EdgeCollapsedRips — Type

EdgeCollapsedRips{I, T} <: AbstractRipsFiltration{I, T}

Perform a sequence of edge collapses on a filtration. This may significantly reduce computation time and does not change the result. The speedup is especially apparent with datasets that have a boundary, and with high-dimensional persistent homology computation.

The drawback is that doing the collapses themselves can be time-consuming. The construction does not require a lot of memory ($\mathcal{O}(n^2)$ for $n$ vertices). This might still be a good choice for large inputs if you are willing to wait but don't have enough memory to compute persistent homology with Rips.

See the reference below for a description of the algorithm.

Constructors

EdgeCollapsedRips(::AbstractRipsFiltration; verbose=false, threshold=nothing ): Collapse a given filtration. Setting verbose shows a progress bar.

EdgeCollapsedRips(::EdgeCollapsedRips; verbose=false, threshold=nothing): Allows changing I or threshold without recomputing.

EdgeCollapsedRips(arg; kwargs...): Use arg and kwargs to construct a Rips filtration and collapse it

EdgeCollapsedRips{I}(arg; kwargs...): Change the index type used to represent simplices. May be necessary for large inputs and high-dimensional computation.

Examples

julia> using Random; Random.seed!(1337);

julia> data = [tuple(rand(6)...) for _ in 1:100];

julia> rips = Rips(data)
Rips{Int64, Float64}(nv=100, sparse=false)

julia> length(Ripserer.edges(rips))
3906

julia> collapsed = EdgeCollapsedRips(data) # or EdgeCollapsedRips(rips)
EdgeCollapsedRips{Int64, Float64}(nv=100)

julia> length(Ripserer.edges(collapsed))
1419

julia> ripserer(rips) == ripserer(collapsed)
true

julia> ripserer(collapsed; dim_max=4)
5-element Array{PersistenceDiagramsBase.PersistenceDiagram,1}:
 100-element 0-dimensional PersistenceDiagram
 75-element 1-dimensional PersistenceDiagram
 37-element 2-dimensional PersistenceDiagram
 14-element 3-dimensional PersistenceDiagram
 1-element 4-dimensional PersistenceDiagram

Reference

Boissonnat, J. D., & Pritam, S. (2019). Edge Collapse and Persistence of Flag Complexes.

Fields

Ripserer.Mod — Type

Mod{M} <: Integer

Mod{M} is the default field used by Ripserer. It is a representation of a finite field $\mathbb{Z}_M$, integers modulo small, prime M. Supports field arithmetic and can be converted to integer with Int.

Its values are not comparable on purpose.

Example

julia> Mod{3}(5)
2 mod 3

julia> Mod{3}(5) + 1
0 mod 3

Persistence Diagrams

See also: PersistenceDiagrams.jl API. For convenience, Ripserer reexports the following:

PersistenceDiagramsBase.birth — Method

birth(interval)

Get the birth time of interval.

PersistenceDiagramsBase.death — Method

death(interval)

Get the death time of interval.

PersistenceDiagramsBase.persistence — Method

persistence(interval)

Get the persistence of interval, which is equal to death - birth.

PersistenceDiagramsBase.representative — Method

representative(interval::PersistenceInterval)

Get the representative (co)cycle attached to interval, if it has one.

PersistenceDiagramsBase.birth_simplex — Method

birth_simplex(interval::PersistenceInterval)

Get the critical birth simplex of interval, if it has one.

PersistenceDiagramsBase.death_simplex — Method

death_simplex(interval::PersistenceInterval)

Get the critical death simplex of interval, if it has one.

Note

An infinite interval's death simplex is nothing.

PersistenceDiagrams.barcode — Function

barcode(diagram)

Plot the barcode plot of persistence diagram or multiple diagrams in a collection. The infinity keyword argument determines where the infinity line is placed. If unset, the function tries to use threshold(diagram), or guess a good position to place the line at.

Experimental Features

Ripserer.reconstruct_cycle — Function

reconstruct_cycle(filtration, interval[, t]; distances=distance_matrix(filtration))

Reconstruct the shortest representative cycle for the first homology group of given interval. The optional argument t sets the time at which the cycle is to be computed. It defaults to interval birth time, which gives a cycle similar to a representative cycle computed from homology. In general, higher times will yield nicer cycles. t can be a simplex or a number.

The optional distances keyword argument can be used to change the distance matrix used for determining edge lengths.

This method uses the representative cocycle to compute the cycle. As such, the interval must include a representative. To get such an interval, run ripserer with the keyword argument reps=true or reps=1.

Warning

This feature is still experimental.

Ripserer.Partition — Module

module Partition

This submodule contains the following partition functions to be used with circular coordinates.

Partition.linear(r, d) = max(r - d, 0.0)

Partition.quadratic(r, d) = (max(r - d, 0.0))^2

Partition.exponential(r, d) = r - d > 0 ? exp(r^2/(d^2 - r^2)) : 0.0

Ripserer.CircularCoordinates — Type

CircularCoordinates

This struct implements Perea's sparse circular coordinates (see reference below).

The idea behind this method is that you pick a subset of your data set (also called landmarks), compute persistent cohomology on that subset, and use the result to construct circular coordinates on the whole data set.

For this to work correctly, you need the following.

A set of landmarks that cover the points well. The default minmax sampling method is

usually a good choice here.

The persistence diagram on the landmarks must have an interval that is persistent enough.

The circular coordinates returned are on the interval [0, 1). If you are looking for angles, make sure to multiply them by 2π.

If you try to compute circular coordinates for a point that is not near any landmark, the result will be missing.

Constructor

CircularCoordinates([::AbstractFiltration, ]points, landmarks; kwargs...)

Arguments

points: a vector of points. The eltype of this vector can be Tuples, SVectors, or similar.

landmarks: can be an integer, a vector of indices, or a vector of points. If set to an integer, it sets the number of landmarks to choose with maxmin sampling. If you are looking for non-sparse circular coordinates, use landmarks=eachindex(points).

out_dim: number of most persistent persistence intervals to use to compute coordinates. If less than out_dim suitable intervals are found, the construction will show a warning and construct CircularCoordinates with a lower out_dim. This warning can be suppressed by passing warn=false.

partition: a function that defines the partition of unity used when determining the coordinates. Should take two arguments, r, the radius of the balls, and d, the distance from the landmark. The partition function should only have support on the ball around the landmark (in other words, it should evaluate to 0 when d ≥ r). If that is not the case, the circular coordinates are not well-defined and the results may not always make sense. See Partition for a small collection of predefined functions.

metric: a metric from Distances.jl. Can only be set if the filtration argument is Rips.

kwargs...: additional keyword arguments passed to ripserer. Note that modulus is set to a random prime between 7 and 79 by default.

Example

julia> data = [(sin(t), cos(t)) for t in range(0, 2π, length=101)[1:end-1]];

julia> cc = CircularCoordinates(data, 1:10:100)
CircularCoordinates(
  out_dim=1,
  radius=(0.9510565162951535,),
  n_landmarks=10,
  partition=linear,
  metric=Distances.Euclidean(0.0),
)

julia> summary(cc(data))
"100×1 Array{Union{Missing, Float64},2}"

julia> summary(cc(data, 1))
"100-element Array{Union{Missing, Float64},1}"

Reference

Perea, J. A. (2020, June). Sparse circular coordinates via principal Z-bundles. In Topological Data Analysis: The Abel Symposium 2018 (Vol. 15, p. 435). Springer Nature.