Simplex Interface

Ripserer.AbstractSimplex — Type

AbstractSimplex{D, T, I} <: AbstractVector{I}

An abstract type for representing simplices. A simplex must have a diameter of type T, which is its birth time. The dimension must be encoded in the type as D and can be accessed by dim.

The simplex is expected to act like an array of indices of type I, but this is not actually needed for the main algorithm.

Interface

AbstractSimplex{D}(::SVector{<:Any, <:Integer}, ::T)

birth(::AbstractSimplex)

index(::AbstractSimplex)

sign(::AbstractSimplex)

coboundary(::Any, ::AbstractSimplex)

boundary(::Any, ::AbstractSimplex)

vertices(::AbstractSimplex)

Base.:-(::AbstractSimplex)

PersistenceDiagrams.dim — Method

dim(::AbstractSimplex)
dim(::Type{<:AbstractSimplex})

Get the dimension of simplex i.e. the value of D.

dim(Simplex{2}((3, 2, 1), 3.2))

# output

2

PersistenceDiagrams.birth — Method

birth(simplex::AbstractSimplex)

Get the birth time of simplex, i.e. the time it appears in a filtration.

birth(Simplex{2}((3, 2, 1), 3.2))

# output

3.2

Ripserer.index — Method

index(simplex::AbstractSimplex)

Get the combinatorial index of the simplex. The index can be any type, but should uniquely identify a simplex. It is also used to break ties when comparing simplices with the same birth time.

index(Simplex{2}((3, 2, 1), 3.2))

# output

1

Base.sign — Method

sign(simplex::AbstractSimplex)

Get the orientation of simplex. Should return -1 or 1.

sign(Simplex{2}((3, 2, 1), 3.2))

# output

+1

Base.:- — Method

-(simplex::AbstractSimplex)

Reverse the simplex orientation.

-Simplex{2}((3, 2, 1), 3.2)

# output

2-dim Simplex(1, 1):
  -[3, 2, 1]

Ripserer.vertices — Method

vertices(simplex::AbstractSimplex{dim, T, I})

Get the vertices of simplex. Returns SVector{length(simplex), I}. When index(simplex) is an interger, a default implementation is provided.

vertices(Simplex{2}((3, 2, 1), 3.2))

# output

3-element StaticArrays.SArray{Tuple{3},Int64,1,3} with indices SOneTo(3):
 3
 2
 1

Ripserer.coboundary — Method

coboundary(filtration, simplex[, Val{all_cofacets}])

Iterate over the coboundary of simplex by decreasing index. Use the filtration to determine the diameters and validity of cofacets.

If all_cofacets is false, only return cofaces with vertices added to the beginning of vertex list. The method with all_cofacets only has to be implemented if the filtration does not overload columns_to_reduce.

Comes with a default implementation.

Warning

If cofacets are not returned in decreasing index, the algorithm will not work correctly.

Example

filtration = Rips([0 1 1 1; 1 0 1 1; 1 1 0 1; 1 1 1 0])

for c in coboundary(filtration, Simplex{1}(2, 1))
    println(c)
end

# output

Simplex{2}(+[4, 3, 1], 1)
Simplex{2}(-[3, 2, 1], 1)

for c in coboundary(filtration, Simplex{1}(2, 1), Val(false))
    println(c)
end

# output

2-dim Simplex(1, 1):
  +[4, 3, 1]

Ripserer.boundary — Method

boundary(filtration, simplex[, Val{all_cofacets}])

Iterate over the boundary of simplex by increasing index. Use the filtration to determine the diameters and validity of cofacets.

Comes with a default implementation.

filtration = Rips([0 1 1 1; 1 0 1 1; 1 1 0 1; 1 1 1 0])

for f in boundary(filtration, Simplex{2}(2, 1))
    println(f)
end

# output

Simplex{2}(+[2, 1], 1)
Simplex{2}(-[4, 1], 1)
Simplex{2}(+[4, 2], 1)